Example data
This is my first project on GitHub. It is intended to learn the basic functions of GitHub, such as uploading files, managing commits, and working with repositories. In this project, I will add some sample data and initial files to understand the structure and organization of a repository. The goal is to develop good practices in version control and open-source project management.